+++ src/parser/block/blockquote.test.ts
@@ -20,17 +20,17 @@ describe('Unit: parser/blockquote', () => {
 
     it('ab', () => {
       assert.deepStrictEqual(inspect(parser('> a')), [['<blockquote>a</blockquote>'], '']);
-      assert.deepStrictEqual(inspect(parser('>  a ')), [['<blockquote>&nbsp;a</blockquote>'], '']);
+      assert.deepStrictEqual(inspect(parser('>  a ')), [['<blockquote>&nbsp;a&nbsp;</blockquote>'], '']);
       assert.deepStrictEqual(inspect(parser('> a\n')), [['<blockquote>a</blockquote>'], '']);
       assert.deepStrictEqual(inspect(parser('> a\nb')), [['<blockquote>a<br>b</blockquote>'], '']);
-      assert.deepStrictEqual(inspect(parser('> a\n b ')), [['<blockquote>a<br>&nbsp;b</blockquote>'], '']);
+      assert.deepStrictEqual(inspect(parser('> a\n b ')), [['<blockquote>a<br>&nbsp;b&nbsp;</blockquote>'], '']);
       assert.deepStrictEqual(inspect(parser('> a\n>')), [['<blockquote>a<br></blockquote>'], '']);
       assert.deepStrictEqual(inspect(parser('> a\n>b')), [['<blockquote>a<br>&gt;b</blockquote>'], '']);
-      assert.deepStrictEqual(inspect(parser('> a\n> b ')), [['<blockquote>a<br>b</blockquote>'], '']);
+      assert.deepStrictEqual(inspect(parser('> a\n> b ')), [['<blockquote>a<br>b&nbsp;</blockquote>'], '']);
       assert.deepStrictEqual(inspect(parser('> a\n>\nb')), [['<blockquote>a<br><br>b</blockquote>'], '']);
-      assert.deepStrictEqual(inspect(parser('> a\n>\n b ')), [['<blockquote>a<br><br>&nbsp;b</blockquote>'], '']);
+      assert.deepStrictEqual(inspect(parser('> a\n>\n b ')), [['<blockquote>a<br><br>&nbsp;b&nbsp;</blockquote>'], '']);
       assert.deepStrictEqual(inspect(parser('> a\n>\n>b')), [['<blockquote>a<br><br>&gt;b</blockquote>'], '']);
-      assert.deepStrictEqual(inspect(parser('> a\n>\n> b ')), [['<blockquote>a<br><br>b</blockquote>'], '']);
+      assert.deepStrictEqual(inspect(parser('> a\n>\n> b ')), [['<blockquote>a<br><br>b&nbsp;</blockquote>'], '']);
       assert.deepStrictEqual(inspect(parser('> a\n\nb')), [['<blockquote>a</blockquote>'], 'b']);
       assert.deepStrictEqual(inspect(parser('> a\nb\nc')), [['<blockquote>a<br>b<br>c</blockquote>'], '']);
       assert.deepStrictEqual(inspect(parser('> a\nb\n\nc')), [['<blockquote>a<br>b</blockquote>'], 'c']);
@@ -46,6 +46,18 @@ describe('Unit: parser/blockquote', () => {
       assert.deepStrictEqual(inspect(parser('>> a\n>>> b\n> c')), [['<blockquote><blockquote>a<blockquote>b</blockquote></blockquote>c</blockquote>'], '']);
       assert.deepStrictEqual(inspect(parser('>> a\n> b\n>>> c')), [['<blockquote><blockquote>a</blockquote>b<blockquote><blockquote>c</blockquote></blockquote></blockquote>'], '']);
       assert.deepStrictEqual(inspect(parser('>>> a\n>> b\n> c')), [['<blockquote><blockquote><blockquote>a</blockquote>b</blockquote>c</blockquote>'], '']);
+    });
+
+    it('markdown', () => {
+      assert.deepStrictEqual(inspect(parser('|')), void 0);
+      assert.deepStrictEqual(inspect(parser('|>')), void 0);
+      assert.deepStrictEqual(inspect(parser('|> ')), void 0);
+      assert.deepStrictEqual(inspect(parser('|> a')), [['<blockquote><p>a</p></blockquote>'], '']);
+      assert.deepStrictEqual(inspect(parser('|>  a \n b  c ')), [['<blockquote><p>a   b  c</p></blockquote>'], '']);
+      assert.deepStrictEqual(inspect(parser('|> a\\\nb')), [['<blockquote><p>a<br>b</p></blockquote>'], '']);
+      assert.deepStrictEqual(inspect(parser('|> **a\nb**')), [['<blockquote><p><strong>a b</strong></p></blockquote>'], '']);
+      assert.deepStrictEqual(inspect(parser('|>> a')), [['<blockquote><blockquote><p>a</p></blockquote></blockquote>'], '']);
+      assert.deepStrictEqual(inspect(parser('|>> a\n> **b**')), [['<blockquote><blockquote><p>a</p></blockquote><p><strong>b</strong></p></blockquote>'], '']);
     });
 
   });